TreeDefinition: allow `Add` a blob by ID
Allow consumers to add a blob to a `TreeDefinition` by specifying only
the object ID and mode.  This lets users build tree entries without
having to create a `Blob` object (and thus load the object itself)
which is beneficial for users building trees with large objects.

This is only useful for Blobs, since they do not need to be fetched from
the object database to be realized within the tree builder.  Users are
not able to add Trees by ID, since we would need to load them (and
cannot, since we are not guaranteed to be instantiated within the
context of a repository).